+++ src/i18n_check/utils.py
@@ -71,8 +71,8 @@ if "global" in config["checks"]:
 # MARK: Invalid Keys
 
 config_invalid_keys_active = config_global_active
-config_invalid_keys_directories_to_skip = config_global_directories_to_skip
-config_invalid_keys_files_to_skip = config_global_files_to_skip
+config_invalid_keys_directories_to_skip = config_global_directories_to_skip.copy()
+config_invalid_keys_files_to_skip = config_global_files_to_skip.copy()
 config_invalid_keys_regex_to_ignore = []
 
 if "invalid-keys" in config["checks"]:
@@ -106,8 +106,8 @@ if "invalid-keys" in config["checks"]:
 # MARK: Non-Existent Keys
 
 config_non_existent_keys_active = config_global_active
-config_non_existent_keys_directories_to_skip = config_global_directories_to_skip
-config_non_existent_keys_files_to_skip = config_global_files_to_skip
+config_non_existent_keys_directories_to_skip = config_global_directories_to_skip.copy()
+config_non_existent_keys_files_to_skip = config_global_files_to_skip.copy()
 
 if "non-existent-keys" in config["checks"]:
     if "active" in config["checks"]["non-existent-keys"]:
@@ -158,8 +158,8 @@ if (
 # MARK: Unused Keys
 
 config_unused_keys_active = config_global_active
-config_unused_keys_directories_to_skip = config_global_directories_to_skip
-config_unused_keys_files_to_skip = config_global_files_to_skip
+config_unused_keys_directories_to_skip = config_global_directories_to_skip.copy()
+config_unused_keys_files_to_skip = config_global_files_to_skip.copy()
 
 if "unused-keys" in config["checks"]:
     if "active" in config["checks"]["unused-keys"]: